+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Changes that are in master but are not yet released:
 
 * Add boolean type as skip option value
+* Add onSend event handler
 
 ## 0.0.1 (October 4th, 2019)